requirements.txt
azure-functions==1.20.0
azure-eventhub==5.12.0
azure-identity==1.16.0
